Add additional compiler warning flags
This commit extends the compiler flags to include many new warnings
which would be normally not checked for.
This should increase code quality as each warning is also now treated as
error.

The flags should enforce best-practices and avoid hard-to-catch bugs

+++ tools/toolsutils.hpp
@@ -94,7 +94,7 @@ namespace toolsutils
 		} else if (bytes == 0) {
 			return true; // EOF reached
 		}
-		if (bytes = ::splice(pipe.get_read_fd(), NULL, to.get(), NULL, bytes, SPLICE_F_MORE | SPLICE_F_MOVE); -1 == bytes) {
+		if (bytes = ::splice(pipe.get_read_fd(), NULL, to.get(), NULL, static_cast<size_t>(bytes), SPLICE_F_MORE | SPLICE_F_MOVE); -1 == bytes) {
 			exit_errno("cannot write data to target file descriptor");
 		}
 		return false;
